Change from (x, y) to Vector2 and from speed to Vector2 (velocityX, velocityY)

+++ src/main/java/inf112/skeleton/model/Globals/Coordinate.java
@@ -1,4 +1,4 @@
 package inf112.skeleton.model.Globals;
 
-public record Coordinate(int x, int y) {
+public record Coordinate(float x, float y) {
 }
